Put client traits behind feature flags
