+++ Cargo.toml
@@ -81,7 +81,7 @@ trussed-derive = { path = "derive" }
 # rand_core = { version = "0.5", features = ["getrandom"] }
 
 [features]
-default = ["default-mechanisms", "default-syscalls", "clients-5"]
+default = ["all-clients", "default-mechanisms", "default-syscalls", "clients-5"]
 serde-extensions = ["trussed-core/serde-extensions"]
 std = []
 verbose-tests = ["littlefs2/ll-assertions"]
@@ -129,9 +129,24 @@ tdes = ["des"]
 totp = ["sha-1"]
 trng = ["sha-1"]
 
+# clients
+all-clients = [
+  "certificate-client",
+  "crypto-client",
+  "counter-client",
+  "filesystem-client",
+  "management-client",
+  "ui-client",
+]
+certificate-client = ["trussed-core/certificate-client"]
+crypto-client = ["trussed-core/crypto-client"]
+counter-client = ["trussed-core/counter-client"]
+filesystem-client = ["trussed-core/filesystem-client"]
+management-client = ["trussed-core/management-client"]
+ui-client = ["trussed-core/ui-client"]
+
 # syscalls
 default-syscalls = ["counter-client", "crypto-client-attest"]
-counter-client = ["trussed-core/counter-client"]
 crypto-client-attest = ["trussed-core/crypto-client-attest"]
 
 clients-1 = []
