+++ core/Cargo.toml
@@ -16,6 +16,12 @@ serde.workspace = true
 serde-indexed = "0.1"
 
 [features]
-crypto-client-attest = []
+certificate-client = []
+crypto-client = []
 counter-client = []
+filesystem-client = []
+management-client = []
+ui-client = []
+
+crypto-client-attest = []
 serde-extensions = []
